move bitvector_typet into bitvector_type.h
This moves the class bitvector_typet from std_types.h into
bitvector_types.h, which is more intuitive.
