+++ unit/analyses/variable-sensitivity/value_set_abstract_object/extremes-go-top.cpp
@@ -7,6 +7,7 @@
 \*******************************************************************/
 
 #include <util/arith_tools.h>
+#include <util/bitvector_types.h>
 #include <util/namespace.h>
 #include <util/symbol_table.h>
